Kroki Integration
This is an intial step towards rendering diagrams in Zulip using textual
descriptions.This feature is brought about by parsing the input text as
a fenced code block.Based on the `lang`, the input is parsed by the
KrokiHandler in `fenced_code.py`.The textual description is
base64encoded by compressing it using the `zlib` library.The request
is then made to the kroki instace with the encoded diagram as a
parameter with the output being the desired svg. The Kroki instance URL is stored in
`prod_settings_template.py`.

Fixes #16825